Created a file name index_session.php
when you set sessions for redis you can verify if the redis is storing the sessions, this can be done using a PHP script. This PHP script implements counter, as each time you reload the page, it prints the number according to it and that will be incremented with the succeeding attempts. Rename your existing index.php temporarily and create a new one with the following code.